Test 45 (more specific)

+++ Amplify/Amplify.swift
@@ -125,7 +125,7 @@ public class TestClass{
         return oldVersion>newVersion-optionalVersion
     }
 }
-typealias UserName = Int
+typealias UserName = Double
 public class SubTestClass: TestClass{
     public override func testAPIBreakage(oldVersion: Int, newVersion: Int, optionalVersion: Int = 10) -> Bool {
         return oldVersion>newVersion+optionalVersion
